### 2020-07-09 #2454 @whedon
[PRE REVIEW]: PySocialForce: A Python Implementation of the Extended Social Force Model for Pedestrian Dynamics
**Submitting author:** @yuxiang-gao (<a href="http://orcid.org/0000-0001-6183-2482">Yuxiang Gao</a>)
**Repository:** <a href="https://github.com/yuxiang-gao/PySocialForce" target ="_blank">https://github.com/yuxiang-gao/PySocialForce</a>
**Version:** v1.0.0
