### 2020-07-13 #2454 @pdebuyl
@moorepants @jackbrookes would you review the article "PySocialForce: A Python Implementation of the Extended Social Force Model for Pedestrian Dynamics" for the Journal of Open Source Software?
