### 2020-07-18 #2454 @pdebuyl
@xuqiancheng would you review the article "PySocialForce: A Python Implementation of the Extended Social Force Model for Pedestrian Dynamics" for the Journal of Open Source Software?
