### 2020-07-19 08cb9dd
:memo: Reduce code listings in the paper
