### 2020-07-21 #2494 @xuqiancheng
@pdebuyl @yuxiang-gao This is a nice implementation of the extended social force model for pedestrian dynamics. It realizes not only the social forces between individuals, but also the social forces for groups. The installation of the code is easy. The tests and exmaple are clear.
From user perspective, I have two major comments regarding the submission:

-  I didn't see the introduction of the parameters in the [default.toml](https://github.com/yuxiang-gao/PySocialForce/blob/master/pysocialforce/utils/default.toml). It's difficult for user to design their own configuration files for simulations. A brief introduction about how is the simulation result changing with these parameters may be useful, especially for researchers who are not familiar with force-based model.

- The result of [scenarios_pedpedspace.py ](https://github.com/yuxiang-gao/PySocialForce/blob/master/tests/scenarios_pedpedspace.py) is about **lane formation** in bi-directional scenario. However it's difficult to observe the phenomenon since the various colors of the pedestrians in the simulation. It would be better to set a same color for pedsetrians moving in same direction. I suggest to add a new parameter to set the color of pedestrians in the visualization part.

Besides, I have three minor comments:

- I install the code according to the [document](https://github.com/yuxiang-gao/PySocialForce#installation). When I install the pip package from PyPI, I need to add [default.toml](https://github.com/yuxiang-gao/PySocialForce/blob/master/pysocialforce/utils/default.toml) into the package manually to run tests succefully. It should be mentioned in the document.

- There are seven *.py files in [tests](https://github.com/yuxiang-gao/PySocialForce/tree/master/tests), three of them are not start with **test**.  I suggest to rename these three files for pytest auto-discovery.

- I didn't see the documents about **Community guidelines**. It should be added.

I would like to ask the authors to consider these suggestions. I will accept this submission after that.
