### 2020-08-04 8f7801c
:memo: Add explanation on the configuration file and the meaning of the parameters
