+++ pysocialforce/utils/default.toml
@@ -8,6 +8,34 @@ max_speed_multiplier = 1.3
 tau = 0.5
 resolution = 10
 
+[goal_attractive_force]
+factor = 1
+
+[ped_repulsive_force]
+factor = 1.5
+v0 = 2.1
+sigma = 0.3
+# fov params
+fov_phi = 100.0
+fov_factor = 0.5 # out of view factor
+
+[space_repulsive_force]
+factor = 1
+u0 = 10
+r = 0.2
+
+[group_coherence_force]
+factor = 3.0
+
+[group_repulsive_force]
+factor = 1.0
+threshold = 0.55
+
+[group_gaze_force]
+factor = 4.0
+# fov params
+fov_phi = 90.0
+
 [desired_force]
 factor = 1.0
 relaxation_time = 0.5
@@ -27,31 +55,3 @@ threshold = 3.0
 
 [along_wall_force]
 
-
-[group_coherence_force]
-factor = 3.0
-
-[group_repulsive_force]
-factor = 1.0
-threshold = 0.55
-
-[group_gaze_force]
-factor = 4.0
-# fov params
-fov_phi = 90.0
-
-[goal_attractive_force]
-factor = 1
-
-[ped_repulsive_force]
-factor = 1.5
-v0 = 2.1
-sigma = 0.3
-# fov params
-fov_phi = 100.0
-fov_factor = 0.5 # out of view factor
-
-[space_repulsive_force]
-factor = 1
-u0 = 10
-r = 0.2
