### 2020-08-04 3dd30b8
Rename tests for auto-discovery,update travis conf
