+++ .travis.yml
@@ -5,7 +5,7 @@ python:
 install:
   - pip install --upgrade pip
   - pip install -e .
-  - pip install --upgrade codecov pylint pytest pytest-cov
+  - pip install --upgrade codecov pylint pytest pytest-cov matplotlib
 script:
   - pylint pysocialforce
   - pytest --cov=pysocialforce tests/
