### 2020-08-04 91c1d69
:memo: Update README for parameter explanation
