### 2020-08-04 5573140
Add Contributing guide
