### 2020-08-04 2c05975
Update contributing guide
