### 2020-08-04 72e6e62
Update travis.yml, pylint error in 3.7
