### 2020-08-04 #2494 @yuxiang-gao
@pdebuyl @xuqiancheng Thank you for the review and I apologize that I couldn't address the comments until now. I have made the following changes to address @xuqiancheng 's concerns:
* For the first major comment and the first minor comment, I have added a more detailed explanation of the configuration file and added a sample file with comments explaning the parameters.
* For the second major comment, I have made the suggested modifications on the visualization part.
* For the rest of the minor comments, I have followed the suggestions by renaming the tests and adding the contributing guides.
