### 2020-08-07 #2494 @ixjlyons
@pdebuyl yes, apologies for the delay. I'm planning to complete my review in the next few days and I'll take the recent updates into account (I had noticed a couple of the same issues already pointed out).
