### 2020-08-10 #2494 @ixjlyons
My review:

PySocialForce is an extension and improvement upon [svenkreiss/socialforce](https://github.com/svenkreiss/socialforce), implementing the social force model with the addition of group forces. The interface for setting up simulations is straightforward and the modeling appears to be easily extended, enabling further research in the field.

I have a few comments below, but otherwise I recommend accepting.

Code/repository comments:

- Chien-Ming Huang is listed as an author but hasn't made public contributions to the source code. Could you state the rationale for authorship?
- Some of the tests appear to generate the images in the README. You might consider migrating these to examples instead (and you could still run them in CI). They appear to be good demonstrations of usage of the library. I was going to suggest moving the files not prefixed with `test_` to the examples folder before the names were changed upon the initial review.
- This project is in what I would consider the grey area of ["substantial scholarly effort"](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). Some code is shared with [svenkreiss/socialforce](https://github.com/svenkreiss/socialforce), but there has been some refactoring and a number of nice features appear to have been added, so I'm thinking these aren't ground for rejection. I just wanted to point this out to the editor -- I noticed the line count guideline, but Python in particular can be quite expressive and dense in terms of lines of code.
- I like the forces model (`forces.py`) - it seems very extensible. Does it make sense to actually unhide `_get_force` to encourage extending the core classes? I'm not very familiar with the field, but this seems like it could open some possibilities for further extending the pedestrian dynamics models significantly. I believe this is referred to in the paper ("adding new social forces").
- Tests fail with installation from PyPI because of the config file (FileNotFoundErrors). I see this has been pointed out in the README since the initial review, but I would recommend packaging the default config file in the sdist. See https://setuptools.readthedocs.io/en/latest/setuptools.html#including-data-files
- Minor comment: there's a typo in the file name generated in `examples/example.py`
- I think there could be a benefit to browsable API documentation. It's not clear to me from the JOSS review guidelines if this is necessary or if general docstring coverage of classes/methods is sufficient (user can get html via pydoc).

Paper comments:

- Should the links to OpenAI Gym, Numba, ROS, and Qt be a citation too/instead?
