### 2020-08-10 #2494 @pdebuyl
Thank you @ixjlyons for the review!

Comments/replies to the review's questions:

1. Installation via pip should proceed successfully.
2. Packaging the files via the "data files" of the package is good practice. We encourage to follow the best practices of the language for the project.
3. The documentation of the API should be available as well. Provided the docstrings are there already, this should be doable with little effort.
4. The documentation is in the README currently. This is ok in itself but it is of course not very flexible.
5. For projects that have a reference paper, it should be cited. Many projects (not all of course) have citation guidelines.
