### 2020-10-25 e5dab5a
Improve: parse requirements in setup.py
