### 2020-10-25 9e93fa1
Improve: Hard code default config
