### 2020-10-27 7021afb
Fix force_vectors initialization
