### 2021-09-09 #2494 @arfon
Due to the lack of input from the author, I'm moving ahead with a rejection here. @ixjlyons, @xuqiancheng, @pdebuyl – many thanks for your efforts here. Sorry this submission didn't make it through to completion.
